requirements.txt
fastapi
uvicorn
groq
tavily-python
pandas
numpy
gradio
gradio_client
python-dotenv
pydantic
